Moved implicit mesh geometry to generation.py
