+++ vispy/geometry/__init__.py
@@ -20,5 +20,5 @@ from .torusknot import TorusKnot  # noqa
 from .calculations import (_calculate_normals, _fast_cross_3d,  # noqa
                            resize)  # noqa
 from .generation import (create_arrow, create_box, create_cone,  # noqa
-                         create_cube, create_cylinder, create_plane,  # noqa
-                         create_sphere)  # noqa
+                         create_cube, create_cylinder, create_implicit_mesh,
+                         create_plane, create_sphere)  # noqa
